pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "planner"
version = "0.1.0"
description = "Plattformuebergreifende Wochenansicht fuer Systemkalender mit PySide6"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.8",
    "holidays>=0.66",
    "pyobjc-framework-Cocoa>=11.0; sys_platform == 'darwin'",
    "pyobjc-framework-EventKit>=11.0; sys_platform == 'darwin'",
    "pywin32>=308; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
]

[project.scripts]
planner = "planner.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/planner"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
